Remove unused Carto code
mapSource and mapStyle are only being used in mapLayer, and mapLayer
isn't being used anywhere else in the file. Removing all three didn't
affect appearance or performance, I think because we're relying
on Leaflet to handle all layers, not Carto. We access Carto to get
the data from its spreadsheets but that's all.